+++ docker/healthcheck.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+curl -s --fail http://localhost:9000/api/status || exit 1